requirements.txt
pandas
apache-airflow
apache-airflow-providers-amazon==6.0.0